importer/requirements.txt
kafka-python
watchdog
jsonschema